LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/semantics/popovers/popover-focus-2-expected.txt

Button1  Button2  Invoker0  Invoker1  Button3  Button4
Invoker  after
Show popover
Toggle popover Other focusable element

FAIL Popover focus navigation assert_equals: Hidden popover should be skipped backwards expected Element node <button id="button1" tabindex="0">Button1</button> but got Element node <button id="button2" tabindex="0">Button2</button>
FAIL Circular reference tab navigation assert_equals: circular reference: Step 5 (backwards) expected Element node <button id="circular4" tabindex="0">after</button> but got Element node <button popovertarget="focus-return1-p" popovertargetacti...
PASS Popover focus returns when popover is hidden by invoker
PASS Popover focus only returns to invoker when focus is within the popover
